🚚 api: [CERTIF] move shareableCertificateSerializer to src/certif/results
