+++ api/src/certification/results/infrastructure/serializers/shareable-certificate-serializer.js
